Also extend test timeout on arch riscv64
The riscv64 builders on Launchpad seem very slow. Attempt to fix the build
failures by doubling the test timeout on riscv64, similar to how it was
done on mipsel.

If ifneg+filter structure as it seems to be the only way to check if
a string is one of multiple options.

https://www.gnu.org/software/make/manual/html_node/Text-Functions.html#Text-Functions
https://man7.org/linux/man-pages/man1/dpkg-architecture.1.html#TERMS