Fix new CMake definitions to correctly use CMAKE_OBJDUMP (Closes: #981652)
Galera-4 fails to cross build from source, because it runs the build
architecture objdump on a host architecture ELF object. This is due to
hard coding the build architecture objdump in galera/src/CMakeLists.txt.
The solution here is to use ${CMAKE_OBJDUMP}, which refers to the
correctly detected objdump for the host architecture. Beyond this, it
also uses ldd, which is bound to fail during cross compilation. Almost
the same effect can be achieved with objdump -x though, so I propose
using it here as well. Please consider applying the attached patch as it
makes galera-4 cross buildable.