MDEV-35045: compile error with deprecated openssl/engine.h
Fedora (41) is deprecating the Openssl engine[1]. The bundled
asio has defines around OPENSSL_NO_ENGINE, but never checks or
sets this.

This adds a check for the openssl/engine.h which might be
included via the openssl-devel-engine package.

Also teh gu_asio_test.cpp needs the err.h header and
not the engine one for the ERR_get_error function.

[1]: https://fedoraproject.org/wiki/Changes/OpensslDeprecateEngine