MDEV-37206 memory leak SSL_COMP_get_compression_methods()
sk_SSL_COMP_zero(SSL_COMP_get_compression_methods()); was only
there for support openssl < 1.0.0. A that has been out of support
for a while we can use the asio provided mechanism.

The the deprecation of compression was ignoring the configuration
for a while so its now just running with cmopression explictly
disabled.