MDEV-35068: MSAN - get_addrinfo
Because of member alignment the struct addrinfo
structure has a 4 byte gap after ai_addrlen (4 byte length)
and ai_addr. The default initializer of this leaves
this uninitialized. MSAN ensure that all members of
structures passed to its intercepted calls, in this
case getaddrinfo, are initialized causing the failure.

This is resolved by ensuring the entire structure
is initialized, both directly and when it is copied.